Created and integrated size data package
